Remove Pyro distributions migrated to PyTorch
