+++ pyro/distributions/__init__.py
@@ -1,7 +1,5 @@
 from __future__ import absolute_import, division, print_function
 
-import warnings
-
 # Notice to Contributors. (@fritzo 2017-12-26)
 #
 # The Pyro team is moving pyro.distributions implementations upstream to
@@ -27,33 +25,17 @@ from pyro.distributions.poisson import Poisson
 from pyro.distributions.random_primitive import RandomPrimitive
 
 # distribution classes with working torch versions in torch.distributions
-try:
-    from pyro.distributions.torch.bernoulli import Bernoulli
-    from pyro.distributions.torch.beta import Beta
-    from pyro.distributions.torch.categorical import Categorical
-    from pyro.distributions.torch.cauchy import Cauchy
-    from pyro.distributions.torch.dirichlet import Dirichlet
-    from pyro.distributions.torch.exponential import Exponential
-    from pyro.distributions.torch.gamma import Gamma
-    from pyro.distributions.torch.multinomial import Multinomial
-    from pyro.distributions.torch.normal import Normal
-    from pyro.distributions.torch.one_hot_categorical import OneHotCategorical
-    from pyro.distributions.torch.uniform import Uniform
-except ImportError:
-    warnings.warn('Using deprecated Pyro 0.1.2 builtin distributions. '
-                  'Please upgrade to latest PyTorch 0.4 prerelease on the master branch.',
-                  DeprecationWarning)
-    from pyro.distributions.bernoulli import Bernoulli
-    from pyro.distributions.beta import Beta
-    from pyro.distributions.categorical import Categorical
-    from pyro.distributions.cauchy import Cauchy
-    from pyro.distributions.dirichlet import Dirichlet
-    from pyro.distributions.exponential import Exponential
-    from pyro.distributions.gamma import Gamma
-    from pyro.distributions.multinomial import Multinomial
-    from pyro.distributions.normal import Normal
-    from pyro.distributions.one_hot_categorical import OneHotCategorical
-    from pyro.distributions.uniform import Uniform
+from pyro.distributions.torch.bernoulli import Bernoulli
+from pyro.distributions.torch.beta import Beta
+from pyro.distributions.torch.categorical import Categorical
+from pyro.distributions.torch.cauchy import Cauchy
+from pyro.distributions.torch.dirichlet import Dirichlet
+from pyro.distributions.torch.exponential import Exponential
+from pyro.distributions.torch.gamma import Gamma
+from pyro.distributions.torch.multinomial import Multinomial
+from pyro.distributions.torch.normal import Normal
+from pyro.distributions.torch.one_hot_categorical import OneHotCategorical
+from pyro.distributions.torch.uniform import Uniform
 
 # function aliases
 bernoulli = RandomPrimitive(Bernoulli)
