+++ pyro/distributions/torch/beta.py
@@ -2,13 +2,23 @@ from __future__ import absolute_import, division, print_function
 
 import torch
 
-from pyro.distributions.beta import Beta as _Beta
 from pyro.distributions.torch_wrapper import TorchDistribution
 from pyro.distributions.util import broadcast_shape, copy_docs_from
 
 
-@copy_docs_from(_Beta)
+@copy_docs_from(TorchDistribution)
 class Beta(TorchDistribution):
+    """
+    Univariate beta distribution parameterized by `alpha` and `beta`.
+
+    This is often used in conjunction with `torch.nn.Softplus` to ensure
+    `alpha` and `beta` parameters are positive.
+
+    :param torch.autograd.Variable alpha: Lower shape parameter.
+        Should be positive.
+    :param torch.autograd.Variable beta: Upper shape parameter.
+        Should be positive.
+    """
     reparameterized = True
 
     def __init__(self, alpha, beta, *args, **kwargs):
